New cli argument - privateKey argument - move all the stuff from src root into utils
